change the bigInteger type to ineger

+++ app/Models/TagJoin.php
@@ -14,9 +14,9 @@ class TagJoin extends Model
     public $table="tag_joins" ;
 
     protected $casts = [
-        'id' => 'BigInteger',
-        'post_id' => 'BigInteger',
-        'tag_id' => 'BigInteger',
+        'id' => 'integer',
+        'post_id' => 'integer',
+        'tag_id' => 'integer',
     ];
     protected $fillable = [
         'id', 'post_id' , 'tag_id'
